Move handlers to new package
